First draft of the Lookup Map
The pattern behind the Ingress Controller is to watch for some or any
cache event (create, update, delete) of certain resources:
- Ingress
- Endpoint
- Service
- Secret
- Pod
- Node

Except for the `update` handler where we have the chance to get a
reflection of the value to check if the item has changed or not, the
result of any handler is to enqueue the object processed: this is going
to trigger the reconciliation loop where the controller lists all the
available Ingress, double check che configuration and in case reload the
HAProxy process.

This is going to happen everytime a Pod, a Secret or an Endpoint is
created although not directly related to an Ingress: we're wasting CPU
cycles and memory without any advantage.

This commit just want to decrease the enqueue rate and enqueue the
related Ingress resources for a specific resource also to allow, for the
future, a smarter reload strategy (tl;dr; don't retrieve all the
ingresses at each reconciliation loop).